dodanie komentarzy i usunięcie nieużywanych usingów

+++ src/GtechDesktop.WPF/Windows/AdminWindows/ProductsManagmentWindow.xaml.cs
@@ -1,18 +1,7 @@
 ﻿using GtechDesktop.WPF.Models;
 using GtechDesktop.WPF.Repositories;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace GtechDesktop.WPF.Windows.AdminWindows
 {
@@ -27,19 +16,22 @@ namespace GtechDesktop.WPF.Windows.AdminWindows
             ProductsListView.ItemsSource = ProductRepository.GetProducts();
         }
 
-        private void EditButtonClick(object sender, RoutedEventArgs e)
+        private void EditButtonClick(object sender, RoutedEventArgs e)//brak edycji
         {
             var product = ((Button)sender).CommandParameter as Product;
             ProductsListView.ItemsSource = ProductRepository.GetProducts();
         }
 
-        private void DeleteButtonClick(object sender, RoutedEventArgs e)
+        private void DeleteButtonClick(object sender, RoutedEventArgs e)//usuwanie produktu
         {
             var product = ((Button)sender).CommandParameter as Product;
-            if (MessageBox.Show("Czy na pewno chcesz usunąć produkt " + product.Name, "G-tech", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+            if(product != null)
             {
-                ProductRepository.DeleteProduct(product);
-                ProductsListView.ItemsSource = ProductRepository.GetProducts();
+                if (MessageBox.Show("Czy na pewno chcesz usunąć produkt " + product.Name, "G-tech", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+                {
+                    ProductRepository.DeleteProduct(product);
+                    ProductsListView.ItemsSource = ProductRepository.GetProducts();
+                }
             }
         }
     }
