+++ src/GtechDesktop.WPF/Windows/AdminWindows/SubcategoriesManagmentWindow.xaml.cs
@@ -1,18 +1,7 @@
 ﻿using GtechDesktop.WPF.Models;
 using GtechDesktop.WPF.Repositories;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace GtechDesktop.WPF.Windows.AdminWindows
 {
@@ -27,21 +16,27 @@ namespace GtechDesktop.WPF.Windows.AdminWindows
             SubcategoriesListView.ItemsSource = SubcategoryRepository.GetSubcategories();
         }
 
-        private void EditButtonClick(object sender, RoutedEventArgs e)
+        private void EditButtonClick(object sender, RoutedEventArgs e)//otwarcie okna do edycji podkategorii
         {
             var subcategory = ((Button)sender).CommandParameter as Subcategory;
-            var addSubcategoryWindow = new AddSubcategoryWindow(subcategory);
-            addSubcategoryWindow.ShowDialog();
-            SubcategoriesListView.ItemsSource = SubcategoryRepository.GetSubcategories();
+            if (subcategory != null ) 
+            { 
+                var addSubcategoryWindow = new AddSubcategoryWindow(subcategory);
+                addSubcategoryWindow.ShowDialog();
+                SubcategoriesListView.ItemsSource = SubcategoryRepository.GetSubcategories();
+            }
         }
 
-        private void DeleteButtonClick(object sender, RoutedEventArgs e)
+        private void DeleteButtonClick(object sender, RoutedEventArgs e)//usuwanie podkategroii
         {
             var subcategory = ((Button)sender).CommandParameter as Subcategory;
-            if (MessageBox.Show("Czy na pewno chcesz usunąć użytkownika " + subcategory.Name, "G-tech", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+            if (subcategory != null )
             {
-                SubcategoryRepository.DeleteSubcategory(subcategory);
-                SubcategoriesListView.ItemsSource = SubcategoryRepository.GetSubcategories();
+                if (MessageBox.Show("Czy na pewno chcesz usunąć użytkownika " + subcategory.Name, "G-tech", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+                {
+                    SubcategoryRepository.DeleteSubcategory(subcategory);
+                    SubcategoriesListView.ItemsSource = SubcategoryRepository.GetSubcategories();
+                }
             }
         }
     }
